Cleared up the last two errors form the test suite.  Both Python2 and Python3 pass now.
